+++ buildSrc/src/main/kotlin/Versions.kt
@@ -13,6 +13,7 @@ const val NOTIFY_VERSION = "1.0.0"
 const val WALLETKIT_VERSION = "1.0.0"
 const val APPKIT_VERSION = "1.0.0"
 const val MODAL_CORE_VERSION = "1.0.0"
+const val YTTRIUM_VERSION = "0.2.1"
 
 //Artifact ids
 const val ANDROID_BOM = "android-bom"
